Reorg library files
Public headers remain in ./include/pdal.

All filters, IO, kernels move from top-level back to ./src. The key difference
between this and the original layout is that the headers for these capabilities
are now treated as local/private headers and are not installed as part of the
public API.

Kernel class (along with KernelInfo, KernelFactory, KernelSupport) move into
the public API.