Add omni kernel
This version of the omni kernel uses PipelineManager. The input reader and
output writer are both required. Optional parameters can be used to
directly specify the reader and writer drivers, else they will be
inferred.

The user may optionally specify multiple filter stages and pass
filter-specific parameters. See apps.rst for an example. The filters will
be applied in the order they are entered.